Fix: admin page skeleton 컴포넌트로 수정
- page.tsx에서는 page만 export 가능
