+++ src/app/(main)/admin/semester/page.tsx
@@ -2,7 +2,7 @@
 
 import { useSuspenseQuery } from '@tanstack/react-query'
 
-import { AdminSemesterSkeleton, SectionWithTitle } from '@/components/feature'
+import { SectionWithTitle } from '@/components/feature'
 import { semesterQueries } from '@/service/api'
 
 import { ActivityAccordion, SemesterList } from './_components'
@@ -23,16 +23,3 @@ const AdminSemesterPage = () => {
 }
 
 export default AdminSemesterPage
-
-export const AdminSemesterPageSkeleton = () => {
-  return (
-    <div className="flex w-full flex-col items-center">
-      <SectionWithTitle title="학기 관리">
-        <AdminSemesterSkeleton />
-      </SectionWithTitle>
-      <SectionWithTitle title="활동 관리">
-        <div />
-      </SectionWithTitle>
-    </div>
-  )
-}
