+++ src/components/feature/skeleton/index.ts
@@ -2,3 +2,4 @@ export { ActivitySkeleton } from './Activity'
 export { AdminSemesterSkeleton, ActivitySemesterSkeleton } from './Semester'
 export { BoardSkeleton } from './Board'
 export { BoardHeroSkeleton } from './BoardHero'
+export { AdminSemesterPageSkeleton } from './SemesterPage'
